Move the selector parser out of `patches/`.
